+++ mapjs/src/core/map-model.js
@@ -174,6 +174,8 @@ module.exports = function MapModel(selectAllTitles, clipboardProvider, defaultRe
 				layoutCompleteOptions = { themeChanged: true };
 			}
 			layoutModel.setLayout(newLayout);
+			// Logging node layout to console to help with debugging:
+			idea_pp(idea)
 			if (!self.isInCollapse) {
 				self.dispatchEvent('layoutChangeComplete', layoutCompleteOptions);
 			}